Add documentation controller tests

+++ LucyCover - Backend/Services/EducationMaterialsService.cs
@@ -1,12 +1,15 @@
 ﻿using AutoMapper;
 using FluentValidation;
 using LucyCover___Backend.Exceptions;
+using LucyCover___Backend.Utility;
 using LucyCover_Database.Repository.IRepository;
 using LucyCover_Model.Database_Entities;
 using LucyCover_Model.Database_Model;
 using LucyCover_Model.DTO_Modeles;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
+using System.Text.RegularExpressions;
 using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace LucyCover___Backend.Services
@@ -16,7 +19,7 @@ namespace LucyCover___Backend.Services
         public List<PatientDTO> GetAssignedPatients(Guid materialId);
         public Task AddNewPatientToMaterial(Guid patientId,Guid materialId);
         public EducationMaterials DownloadMaterial(Guid educationMaterialId);
-        public Task UpsertEducationMaterial(EducationMaterialDTO educationMaterialDTO);
+        public Task<string> UpsertEducationMaterial(EducationMaterialDTO educationMaterialDTO);
         public void DeleteMaterial(Guid materialId);
     }
     public class EducationMaterialsService : IEducationMaterialsService
@@ -26,13 +29,16 @@ namespace LucyCover___Backend.Services
         private IWebHostEnvironment _webHostEnvironment { get; set; }
         private IEmailService _emailService { get;set;}
         private readonly Guid _loggedUser;
-        public EducationMaterialsService(IUnitOfWork unitOfWork,IMapper mapper,IWebHostEnvironment webHostEnvironment, IEmailService emailService,IAuthenticationService authenticationService)
+        private readonly string _uploadsDirecotry;
+        public EducationMaterialsService(IUnitOfWork unitOfWork,IMapper mapper,IWebHostEnvironment webHostEnvironment, 
+                                         IEmailService emailService,IAuthenticationService authenticationService, IOptions<FileServerConfig> fileServerOptions)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
             _webHostEnvironment = webHostEnvironment;
             _emailService = emailService;
             _loggedUser = authenticationService.GetCurrentUserId();
+            _uploadsDirecotry = fileServerOptions.Value.Directory;
         }
 
         public List<EducationMaterials> GetAll() 
@@ -43,6 +49,9 @@ namespace LucyCover___Backend.Services
 
         public List<PatientDTO> GetAssignedPatients(Guid materialId)
         {
+            var test = _unitOfWork.educationMaterials.GetAll();
+            if(!_unitOfWork.educationMaterials.Any(m => m.Id == materialId)) throw new EntityNotFoundException("Material with provided id does not exist");
+
             List<EducationMaterialsAssignedPatients> list = _unitOfWork.educationMaterialsAssignedPatients.GetAll(entity => entity.educationMaterialsId== materialId,includeProperties:"patient").ToList();
             return _mapper.Map<List<PatientDTO>>(list);
         }
@@ -56,14 +65,13 @@ namespace LucyCover___Backend.Services
 
             EducationMaterials material = _unitOfWork.educationMaterials.GetFirstOfDefault(material => material.Id== materialId);
 
-            if(material is null || material.userId != _loggedUser)
+            if(material is null)
             {
                 throw new EntityNotFoundException("Material with that id doesn't exist or you have not permissions to it");
             }
-
-            if(patient.email.Trim().Length<=0 || patient.email is null) 
+            else if (material.userId != _loggedUser)
             {
-                throw new EmailValidationException("Patient does not have email assigned"); 
+                throw new UnauthorizedAccessException("You can't access to this resources");
             }
 
             bool materialAssignmentPatient = _unitOfWork.educationMaterialsAssignedPatients.Any(entity => entity.patientId == patientId && entity.educationMaterialsId == materialId);
@@ -84,29 +92,36 @@ namespace LucyCover___Backend.Services
 
         }
 
-        public async Task UpsertEducationMaterial(EducationMaterialDTO educationMaterialDTO)
+        public async Task<string> UpsertEducationMaterial(EducationMaterialDTO educationMaterialDTO)
         {
+            string filePath;
+
             if(educationMaterialDTO.id == Guid.Empty)
             {
-                await AddNewMaterial(educationMaterialDTO);
+               filePath = await AddNewMaterial(educationMaterialDTO);
             }
             else 
             {
-                await EditMaterial(educationMaterialDTO);
+               filePath = await EditMaterial(educationMaterialDTO);
             }
+
+            return filePath;
         }
 
         public EducationMaterials DownloadMaterial(Guid educationMaterialId)
         {
-            EducationMaterials material = _unitOfWork.educationMaterials.GetFirstOfDefault(m => m.Id == educationMaterialId && m.userId == _loggedUser);
+            EducationMaterials material = _unitOfWork.educationMaterials.GetFirstOfDefault(m => m.Id == educationMaterialId);
 
             if(material is null)
             {
                 throw new EntityNotFoundException("Material with that id does not exist or you have not permissions to edit it");
             }
 
+            if(material.userId != _loggedUser) throw new UnauthorizedAccessException();
+
             if(!File.Exists(material.filePath)) 
             {
+                _unitOfWork.educationMaterials.Remove(material);
                 throw new FileNotFoundException();
             }
 
@@ -115,14 +130,16 @@ namespace LucyCover___Backend.Services
 
         public void DeleteMaterial(Guid materialId) 
         {
-            EducationMaterials material = _unitOfWork.educationMaterials.GetFirstOfDefault(m => m.Id == materialId && m.userId == _loggedUser);
+            EducationMaterials material = _unitOfWork.educationMaterials.GetFirstOfDefault(m => m.Id == materialId);
 
             if(material is null)
             {
                 throw new EntityNotFoundException("Material with that id does not exist or you have not permissions to edit it");
             }
 
-            if(!File.Exists(material.filePath))
+            if(material.userId != _loggedUser) throw new UnauthorizedAccessException("Material is not assigned to current user");
+
+            if (!File.Exists(material.filePath))
             {
                 throw new FileNotFoundException();
             }
@@ -132,7 +149,7 @@ namespace LucyCover___Backend.Services
             _unitOfWork.Save();
         }
 
-        private async Task AddNewMaterial(EducationMaterialDTO educationMaterialDTO)
+        private async Task<string> AddNewMaterial(EducationMaterialDTO educationMaterialDTO)
         {    
                 string filePath = await SaveFile(educationMaterialDTO.file);
 
@@ -154,21 +171,25 @@ namespace LucyCover___Backend.Services
                 {
                     throw new FileNotFoundException("File does not save in system");
                 }
-
+                return filePath;
         }
 
-        private async Task EditMaterial(EducationMaterialDTO educationMaterialDTO)
+        private async Task<string> EditMaterial(EducationMaterialDTO educationMaterialDTO)
         {
-            EducationMaterials material = _unitOfWork.educationMaterials.GetFirstOfDefault(m => m.Id == educationMaterialDTO.id && m.userId == _loggedUser);
+            EducationMaterials material = _unitOfWork.educationMaterials.GetFirstOfDefault(m => m.Id == educationMaterialDTO.id);
 
             if(material is null)
             {
                 throw new EntityNotFoundException("Material with that id does not exist or you have not permissions to edit it");
             }
 
+            if(material.userId != _loggedUser) throw new UnauthorizedAccessException("Access to this material was forbidden !");
+
+            var filePath = material.filePath;
+
             if(educationMaterialDTO.file != null) 
             {
-                string filePath = await SaveFile(educationMaterialDTO.file);
+                filePath = await SaveFile(educationMaterialDTO.file);
                 File.Delete(material.filePath);
                 material.fileName = educationMaterialDTO.file.FileName;
                 material.filePath = filePath;
@@ -180,13 +201,14 @@ namespace LucyCover___Backend.Services
             }
             _unitOfWork.educationMaterials.Update(material);
             _unitOfWork.Save();
+            return filePath;
         }
 
         private async Task<string> SaveFile(IFormFile file)
         {
             if(file is not null && file.Length > 0) 
             {
-                string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath,"Uploads");
+                string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath,_uploadsDirecotry);
 
                 if(!Directory.Exists(uploadsFolder))
                 {
